+++ packages/wds/src/components/date-calendar/helpers.ts
@@ -1,0 +1,175 @@
+import dayjs from 'dayjs';
+
+import type { Dayjs } from 'dayjs';
+import type { RefObject } from 'react';
+import type { DateType, ViewType } from './types';
+
+type DefaultDateHelperParams = {
+  max: DateType;
+  min: DateType;
+  value: DateType;
+  timezone: string | undefined;
+};
+
+export const isValidDate = (date: DateType): date is Date | string => {
+  if (date === undefined || date === null) {
+    return false;
+  }
+
+  return dayjs(date).isValid();
+};
+
+export const isDisabledDate = ({
+  min,
+  max,
+  value,
+  timezone,
+}: DefaultDateHelperParams) => {
+  if (
+    isValidDate(min) &&
+    dayjsTimezone(dayjs(min), timezone).isAfter(
+      dayjsTimezone(dayjs(value), timezone),
+    )
+  ) {
+    return true;
+  }
+
+  if (
+    isValidDate(max) &&
+    dayjsTimezone(dayjs(max), timezone).isBefore(
+      dayjsTimezone(dayjs(value), timezone),
+    )
+  ) {
+    return true;
+  }
+
+  return false;
+};
+
+export const dateTypeToDateObject = (
+  v: DateType | Dayjs,
+  timezone: string | undefined,
+) =>
+  typeof v === 'object'
+    ? new Date(dayjsTimezone(dayjs(v), timezone).toISOString())
+    : new Date(dayjsTimezone(dayjs(new Date(v!)), timezone).toISOString());
+
+export const dayjsTimezone = (v: Dayjs, timezone: string | undefined) => {
+  if (timezone?.toLowerCase() === 'utc') {
+    return v.utc();
+  }
+
+  return v.tz(timezone);
+};
+
+export const findClosestEnableDate = ({
+  min,
+  max,
+  value,
+  timezone,
+}: DefaultDateHelperParams) => {
+  if (
+    isValidDate(min) &&
+    dayjsTimezone(dayjs(min), timezone).isAfter(
+      dayjsTimezone(dayjs(value), timezone),
+    )
+  ) {
+    return dateTypeToDateObject(min, timezone);
+  }
+
+  if (
+    isValidDate(max) &&
+    dayjsTimezone(dayjs(max), timezone).isBefore(
+      dayjsTimezone(dayjs(value), timezone),
+    )
+  ) {
+    return dateTypeToDateObject(max, timezone);
+  }
+
+  return dateTypeToDateObject(value, timezone);
+};
+
+export const getWeekdays = (locale?: string) => {
+  return new Array(7).fill(0).map((_, i) => ({
+    narrow: new Intl.DateTimeFormat(locale, {
+      weekday: 'narrow',
+    }).format(dayjs().weekday(i).toDate()),
+    long: new Intl.DateTimeFormat(locale, {
+      weekday: 'long',
+    }).format(dayjs().weekday(i).toDate()),
+    short: new Intl.DateTimeFormat(locale, {
+      weekday: 'short',
+    }).format(dayjs().weekday(i).toDate()),
+  }));
+};
+
+export const getMonths = (locale?: string) => {
+  return new Array(12).fill(0).map((_, i) => ({
+    '2-digit': new Intl.DateTimeFormat(locale, {
+      month: '2-digit',
+    }).format(dayjs().month(i).toDate()),
+    long: new Intl.DateTimeFormat(locale, {
+      month: 'long',
+    }).format(dayjs().month(i).toDate()),
+    narrow: new Intl.DateTimeFormat(locale, {
+      month: 'narrow',
+    }).format(dayjs().month(i).toDate()),
+    short: new Intl.DateTimeFormat(locale, {
+      month: 'short',
+    }).format(dayjs().month(i).toDate()),
+    numeric: new Intl.DateTimeFormat(locale, {
+      month: 'numeric',
+    }).format(dayjs().month(i).toDate()),
+  }));
+};
+
+export const getMeridiem = (locale?: string) => {
+  return new Array(2).fill(0).map((_, i) => ({
+    lower:
+      new Intl.DateTimeFormat(locale, {
+        hour: 'numeric',
+        hour12: true,
+      })
+        .formatToParts(
+          dayjs()
+            .hour(i * 12)
+            .toDate(),
+        )
+        .find((v) => v.type === 'dayPeriod')?.value || 'a',
+    upper:
+      new Intl.DateTimeFormat(locale, {
+        hour: 'numeric',
+        hour12: true,
+      })
+        .formatToParts(
+          dayjs()
+            .hour(i * 12)
+            .toDate(),
+        )
+        .find((v) => v.type === 'dayPeriod')
+        ?.value.toUpperCase() || 'A',
+  }));
+};
+
+export const focusDate = (
+  type: ViewType,
+  value: number,
+  containerRef: RefObject<HTMLDivElement>,
+) => {
+  switch (type) {
+    case 'year':
+      return containerRef.current
+        ?.querySelector<HTMLDivElement>(`[data-year='${value}']`)
+        ?.focus();
+    case 'month':
+      return containerRef.current
+        ?.querySelector<HTMLDivElement>(`[data-month='${value}']`)
+        ?.focus();
+    case 'day':
+      return containerRef.current
+        ?.querySelector<HTMLDivElement>(
+          `[data-day='${value}']:not([aria-disabled='true'])[data-other-month='false']`,
+        )
+        ?.focus();
+  }
+};